Don't export private normalize & refactor tests
