+++ ember-can/rollup.config.mjs
@@ -22,21 +22,15 @@ export default {
     // See https://github.com/embroider-build/embroider/blob/main/docs/v2-faq.md#how-can-i-define-the-public-exports-of-my-addon
     addon.publicEntrypoints([
       'index.js',
-      'computed.js',
       'ability.ts',
       'helpers/**/*.js',
       'services/**/*.js',
-      'utils/**/*.js',
     ]),
 
     // These are the modules that should get reexported into the traditional
     // "app" tree. Things in here should also be in publicEntrypoints above, but
     // not everything in publicEntrypoints necessarily needs to go here.
-    addon.appReexports([
-      'helpers/**/*.js',
-      'services/**/*.js',
-      'utils/**/*.js',
-    ]),
+    addon.appReexports(['helpers/**/*.js', 'services/**/*.js']),
 
     // Follow the V2 Addon rules about dependencies. Your code can import from
     // `dependencies` and `peerDependencies` as well as standard Ember-provided
