+++ ember-can/src/-private/normalize.ts
